working through snapshot loading on wasm

+++ crates/runmat-snapshot/src/format.rs
@@ -100,22 +100,22 @@ pub struct PerformanceMetrics {
     pub creation_time: Duration,
 
     /// Number of builtins captured
-    pub builtin_count: usize,
+    pub builtin_count: u64,
 
     /// HIR cache entries
-    pub hir_cache_entries: usize,
+    pub hir_cache_entries: u64,
 
     /// Bytecode cache entries
-    pub bytecode_cache_entries: usize,
+    pub bytecode_cache_entries: u64,
 
     /// Total uncompressed size
-    pub uncompressed_size: usize,
+    pub uncompressed_size: u64,
 
     /// Compression ratio achieved
     pub compression_ratio: f64,
 
     /// Memory usage during creation
-    pub peak_memory_usage: usize,
+    pub peak_memory_usage: u64,
 }
 
 /// Target platform information
@@ -154,10 +154,10 @@ pub struct DataSectionInfo {
     pub compression: CompressionInfo,
 
     /// Uncompressed data size
-    pub uncompressed_size: usize,
+    pub uncompressed_size: u64,
 
     /// Compressed data size
-    pub compressed_size: usize,
+    pub compressed_size: u64,
 
     /// Data section offset in file
     pub data_offset: u64,
@@ -506,11 +506,11 @@ impl SnapshotFormat {
 
     /// Get total file size
     pub fn total_size(&self) -> usize {
-        let header_size = bincode::serialized_size(&self.header).unwrap_or(0) as usize;
-        let data_size = self.data.len();
-        let checksum_size = self.checksum.as_ref().map_or(0, |c| c.len());
+        let header_size = bincode::serialized_size(&self.header).unwrap_or(0) as u64;
+        let data_size = self.data.len() as u64;
+        let checksum_size = self.checksum.as_ref().map_or(0, |c| c.len()) as u64;
 
-        header_size + data_size + checksum_size
+        (header_size + data_size + checksum_size) as usize
     }
 }
